Update OAuth2 dependency #518
- package "crypto/rand" is being imported twice
- Remove unused reference
- Upgrade to golang/x/oauth2

Resolves #518

+++ doc.go
@@ -438,10 +438,10 @@ type userContext struct {
 
 // tokenResponse
 type tokenResponse struct {
-	TokenType    string `json:"token_type"`
-	AccessToken  string `json:"access_token"`
-	IDToken      string `json:"id_token"`
-	RefreshToken string `json:"refresh_token,omitempty"`
-	ExpiresIn    int    `json:"expires_in"`
-	Scope        string `json:"scope,omitempty"`
+	TokenType    string  `json:"token_type"`
+	AccessToken  string  `json:"access_token"`
+	IDToken      string  `json:"id_token"`
+	RefreshToken string  `json:"refresh_token,omitempty"`
+	ExpiresIn    float64 `json:"expires_in"`
+	Scope        string  `json:"scope,omitempty"`
 }
